+++ middlewares/selectAndAuthzInstructorQuestion.sql
@@ -1,23 +1,23 @@
 -- BLOCK select_and_auth
-WITH error_count AS (
-    SELECT count(*) AS open_error_count
-    FROM errors AS e
+WITH issue_count AS (
+    SELECT count(*) AS open_issue_count
+    FROM issues AS i
     WHERE
-        e.question_id = $question_id
-        AND e.course_caused
-        AND e.open
+        i.question_id = $question_id
+        AND i.course_caused
+        AND i.open
 )
 SELECT
     to_json(q) AS question,
     to_json(top) AS topic,
     tags_for_question(q.id) AS tags,
     assessments_format_for_question(q.id, ci.id) AS assessments,
-    error_count.open_error_count
+    issue_count.open_issue_count
 FROM
     questions as q
     JOIN topics as top ON (top.id = q.topic_id),
     course_instances AS ci,
-    error_count
+    issue_count
 WHERE
     q.id = $question_id
     AND ci.id = $course_instance_id