+++ pages/studentInstanceQuestionHomework/studentInstanceQuestionHomework.js
@@ -49,14 +49,55 @@ function processSubmission(req, res, callback) {
     });
 }
 
+function processIssue(req, res, callback) {
+    if (!res.locals.assessment.allow_issue_reporting) return callback(new Error('Issue reporting not permitted for this assessment'));
+    const description = req.body.description;
+    if (!_.isString(description) || description.length == 0) {
+        return callback(new Error('A description of the issue must be provided'));
+    }
+
+    const variant_id = req.body.__variant_id;
+    sqldb.callOneRow('variants_ensure_instance_question', [variant_id, res.locals.instance_question.id], (err, _result) => {
+        if (ERR(err, callback)) return;
+
+        const course_data = _.pick(res.locals, ['variant', 'instance_question',
+                                                'question', 'assessment_instance',
+                                                'assessment', 'course_instance', 'course']);
+        const params = [
+            variant_id,
+            description, // student message
+            'student-reported issue', // instructor message
+            true, // manually_reported
+            true, // course_caused
+            course_data,
+            {}, // system_data
+            res.locals.authn_user.user_id,
+        ];
+        sqldb.call('issues_insert_for_variant', params, (err) => {
+            if (ERR(err, callback)) return;
+            callback(null, variant_id);
+        });
+    });
+}
+
 router.post('/', function(req, res, next) {
     if (res.locals.assessment.type !== 'Homework') return next();
     if (!res.locals.authz_result.authorized_edit) return next(error.make(403, 'Not authorized', res.locals));
-    processSubmission(req, res, function(err, variant_id) {
-        if (ERR(err, next)) return;
-        res.redirect(res.locals.urlPrefix + '/instance_question/' + res.locals.instance_question.id
-                     + '/?variant_id=' + variant_id);
-    });
+    if (req.body.__action == 'grade' || req.body.__action == 'save') {
+        processSubmission(req, res, function(err, variant_id) {
+            if (ERR(err, next)) return;
+            res.redirect(res.locals.urlPrefix + '/instance_question/' + res.locals.instance_question.id
+                         + '/?variant_id=' + variant_id);
+        });
+    } else if (req.body.__action == 'report_issue') {
+        processIssue(req, res, function(err, variant_id) {
+            if (ERR(err, next)) return;
+            res.redirect(res.locals.urlPrefix + '/instance_question/' + res.locals.instance_question.id
+                         + '/?variant_id=' + variant_id);
+        });
+    } else {
+        next(error.make(400, 'unknown __action: ' + req.body.__action, {locals: res.locals, body: req.body}));
+    }
 });
 
 router.get('/', function(req, res, next) {