+++ sprocs/issues_insert_for_variant.sql
@@ -1,10 +1,12 @@
-DROP FUNCTION IF EXISTS errors_insert_for_variant(bigint,text,boolean,jsonb,jsonb,bigint);
+DROP FUNCTION IF EXISTS issues_insert_for_variant(bigint,text,boolean,jsonb,jsonb,bigint);
+DROP FUNCTION IF EXISTS issues_insert_for_variant(bigint,text,boolean,boolean,jsonb,jsonb,bigint);
 
 CREATE OR REPLACE FUNCTION
-    errors_insert_for_variant(
+    issues_insert_for_variant(
         variant_id bigint,
         student_message text,
         instructor_message text,
+        manually_reported boolean,
         course_caused boolean,
         course_data jsonb,
         system_data jsonb,
@@ -16,13 +18,14 @@ DECLARE
     course_instance_id bigint;
     question_id bigint;
     assessment_id bigint;
+    instance_question_id bigint;
     user_id bigint;
     display_id text;
 BEGIN
     SELECT
-        c.id,      ci.id,              q.id,        a.id,        u.user_id
+        c.id,      ci.id,              q.id,        a.id,                         iq.id, u.user_id
     INTO
-        course_id, course_instance_id, question_id, assessment_id, user_id
+        course_id, course_instance_id, question_id, assessment_id, instance_question_id,   user_id
     FROM
         variants AS v
         JOIN questions AS q ON (q.id = v.question_id)
@@ -38,13 +41,15 @@ BEGIN
 
     IF NOT FOUND THEN RAISE EXCEPTION 'invalid variant_id'; END IF;
 
-    display_id := errors_generate_display_id();
+    display_id := issues_generate_display_id();
 
-    INSERT INTO errors
+    INSERT INTO issues
         (display_id, student_message, instructor_message, course_caused, course_data, system_data, authn_user_id,
-        course_id, course_instance_id, question_id, assessment_id, user_id, variant_id)
+        instance_question_id, course_id, course_instance_id, question_id, assessment_id, user_id, variant_id,
+        manually_reported)
     VALUES
         (display_id, student_message, instructor_message, course_caused, course_data, system_data, authn_user_id,
-        course_id, course_instance_id, question_id, assessment_id, user_id, variant_id);
+        instance_question_id, course_id, course_instance_id, question_id, assessment_id, user_id, variant_id,
+        manually_reported);
 END;
 $$ LANGUAGE plpgsql VOLATILE;