+++ sprocs/issues_update_open_all.sql
@@ -1,5 +1,5 @@
 CREATE OR REPLACE FUNCTION
-    errors_update_open_all (
+    issues_update_open_all (
         new_open boolean,
         course_id bigint,
         authn_user_id bigint
@@ -8,11 +8,11 @@ AS $$
 DECLARE
     old_open boolean;
 BEGIN
-    WITH updated_errors AS (
-        UPDATE errors AS e
+    WITH updated_issues AS (
+        UPDATE issues AS e
         SET open = new_open
         WHERE
-            e.course_id = errors_update_open_all.course_id
+            e.course_id = issues_update_open_all.course_id
             AND e.course_caused
             AND e.open IS DISTINCT FROM new_open
         RETURNING e.id, e.open
@@ -23,9 +23,9 @@ BEGIN
         parameters, new_state)
     SELECT
         authn_user_id, course_id,
-        'errors',  'open', e.id, 'update',
+        'issues',  'open', e.id, 'update',
         jsonb_build_object('course_id', course_id),
         jsonb_build_object('open', e.open)
-    FROM updated_errors AS e;
+    FROM updated_issues AS e;
 END;
 $$ LANGUAGE plpgsql VOLATILE;